% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_gis.R
\name{export_gis}
\alias{export_gis}
\title{Exports a 'MultiLand' object as GIS data}
\usage{
export_gis(
  x,
  raster = NULL,
  points = NULL,
  radii = NULL,
  ext_raster = NULL,
  dir = getwd(),
  filenames = c("points", "buffers", "lsm_raster", "ext_raster"),
  overwrite = FALSE,
  filetype = c("ESRI Shapefile", "GTiff"),
  gdal = c("COMPRESS=DEFLATE", "PREDICTOR=2", "ZLEVEL=9"),
  ...
)
}
\arguments{
\item{x}{An object of class 'MultiLand' generated with \code{\link[=mland]{mland()}}.}

\item{raster, ext_raster}{Numeric. The rasterlayers to be exported.}

\item{points}{Numeric or character vector of points to be processed. See Details.}

\item{radii}{Numeric vector of radii to be processed.}

\item{dir}{Character. Name of the target directory where files will be exported. Provide a
string with an existing folder, otherwise the function will create it.}

\item{filenames}{Character vector with the root of the file names to be exported, for
points, buffers, main rasterlayers and extra rasters, respectively. Default is
c("points", "buffers", "lsm_raster", "ext_raster").}

\item{overwrite}{Logical. Whether to overwrite (TRUE) or not (FALSE, default) existing files
with the same names.}

\item{filetype}{Character. File format expressed as GDAL driver names
(\href{https://gdal.org/drivers/vector/index.html}{Vector Drivers},
\href{https://gdal.org/drivers/raster/index.html}{Raster drivers}). Default are
c("ESRI Shapefile", "GTiff") for points/rasters and raster intersections, respectively.}

\item{gdal}{GDAL driver specific datasource creation options. See the GDAL documentation. With the
\href{https://gdal.org/drivers/raster/gtiff.html}{GeoTiff file format}, \code{\link[=export_gis]{export_gis()}} uses the
following compression options: c("COMPRESS=DEFLATE", "PREDICTOR=2", "ZLEVEL=9").}

\item{...}{Other arguments passed to \link[terra:writeRaster]{terra::writeRaster}.}
}
\value{
No return of an R object. GIS data from a 'MultiLand' object is exported.
}
\description{
Exports points, buffers and intersections between buffers and rasterlayers, as shapefiles and raster
files.
}
\details{
If argument \code{points} is a character vector,
\code{\link[=export_gis]{export_gis()}} will assume that the 'MultiLand' object inputted in argument \code{x} was created with
\code{site_ref = TRUE}. This is, there is an attribute in points layer data with the names for
each individual point. Therefore, the inputted values in argument \code{points} will be taken as these
identification names. Otherwise, if a numeric vector is declared, the inputted values
will be taken as the automatically generated point ids (created when running \code{\link[=mland]{mland()}}).
}
\examples{
\dontrun{
# Loads a 'MultiLand' object
ernesdesign <- system.file("extdata", "ernesdesign.zip", package = "multilandr")
ernesdesign <- load_mland(ernesdesign)

# Exports as GIS data
export_gis(ernesdesign, dir = "ernesdesign")
}
}
\seealso{
\code{\link[=mland]{mland()}}
}
