% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mland.R
\name{load_mland}
\alias{load_mland}
\title{Load 'MultiLand' or 'MultiLandMetrics' object}
\usage{
load_mland(path, ...)
}
\arguments{
\item{path}{A string depicting the path to a zip file or to a folder, to load objects of class
'MultiLand', or to a RDS file to load objects of class 'MultiLandMetrics'.}

\item{...}{Other parameters passed to \code{\link[=readRDS]{readRDS()}} when trying to load an object of class
'MultiLandMetrics'.}
}
\value{
A 'MultiLand' or a 'MultiLandMetrics' object.
}
\description{
Imports a zip file or a folder into an object of class 'MultiLand' that was previously
saved with \code{\link[=save_mland]{save_mland()}}. Alternatively, loads to the environment an RDS object depicting a
'MultiLandMetrics' object.
}
\examples{
# Loads mland object from a zip file, previously created with save_mland()
mland_obj <- system.file("extdata", "ernesdesign.zip", package = "multilandr")
ernesdesign <- load_mland(mland_obj)

# Loads a MultiLandMetrics object previously generated with metrics() and exported as a
# RDS object with save_mland() or saveRDS()

mlm_obj <- system.file("extdata", "ed_metrics.rds", package = "multilandr")
ed_metrics <- load_mland(mlm_obj)
}
\seealso{
\code{\link[=save_mland]{save_mland()}}, \code{\link[=mland]{mland()}}, \code{\link[=metrics]{metrics()}}
}
