% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multilandr_data}
\alias{multilandr_data}
\title{multilandr_data}
\format{
An object of class \code{character} of length 5.
}
\usage{
multilandr_data
}
\description{
A character vector containing the links to external data (.tif and .gpkg files) to be used in examples.
}
\keyword{datasets}
