% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,MultiLand-method}
\alias{show,MultiLand-method}
\title{Show 'MultiLand' object}
\usage{
\S4method{show}{MultiLand}(object)
}
\arguments{
\item{object}{Prints relevant information about a 'MultiLand' object.}
}
\value{
No return value, called for side effects.
}
\description{
Show 'MultiLand' object
}
