% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,MultiLandMetrics-method}
\alias{show,MultiLandMetrics-method}
\title{Show 'MultiLandMetrics' object}
\usage{
\S4method{show}{MultiLandMetrics}(object)
}
\arguments{
\item{object}{Prints relevant information about a 'MultiLandMetrics' object.}
}
\description{
Show 'MultiLandMetrics' object
}
